+++ MySensors_Node_Sensor_Dht22.h
@@ -30,7 +30,7 @@ class MySensors_Node_Sensor_Dht22 : public MySensors_Node_Sensor {
                                  uint8_t  humidity_id     = 11,
                                  String   description     = "DHT",
                                  uint8_t  pin             = 4,      // Pin for dht22 device
-                                 uint16_t interval        = 120000, // How often to read and send in ms
+                                 uint32_t interval        = 120000, // How often to read and send in ms
                                  int16_t  offset          = 0       // Fixed offset of temperature
                                );
     ~MySensors_Node_Sensor_Dht22();
@@ -44,7 +44,7 @@ class MySensors_Node_Sensor_Dht22 : public MySensors_Node_Sensor {
   private:
     uint8_t  _humidity_id;
     uint8_t  _pin;
-    uint16_t _interval;
+    uint32_t _interval;
     int16_t  _offset;
 
     uint32_t _next_read;
